% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gitignore.R
\name{build_gitignore}
\alias{build_gitignore}
\title{Build .gitignore}
\usage{
build_gitignore(dir = getwd())
}
\arguments{
\item{dir}{The location to create the .gitignore.  Defaults to working directory.}
}
\description{
Generates a .gitignore file and adds it to the repo.
}
\examples{
build_gitignore()
}
