% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{col_sum_na}
\alias{col_sum_na}
\title{Count the NAs in each column}
\usage{
col_sum_na(data)
}
\value{
NA count for each
}
\description{
\lifecycle{maturing}
Count all the NAs in each column of a data frame
}
