% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{common_vars}
\alias{common_vars}
\title{common_vars}
\usage{
common_vars(...)
}
\arguments{
\item{...}{Bare, unquoted tibble object names.}
}
\description{
\lifecycle{experimental}
Shows the variable names that are in common between two or more tibbles.
}
