\name{efd}
\alias{efd}

\title{Fall Enrollment: EFD}

\description{
This file contains data on the total entering class, first-year retention rates
and the student-to-faculty ratio.
}
\section{Total Entering Class}{
The total entering class is the number of
incoming students (students enrolling for the first time in a postsecondary
institution versus students transferring in from another postsecondary
institution) at an institution.

This measure permits the calculation of the
graduation rate cohort as a proportion of the total entering student body.
}
\section{Retention Rate}{
The
first-year retention rate data measure the percentage of first-year students
who had persisted in or completed their educational program a year later (began
in fall of the previous year and persisted or completed in fall of the current year).
}

\section{Exclusions}{
Institutions can exclude the number of students from the first-time cohort who left the institution for
any of the following reasons:
  \itemize{
    \item died or were totally and permanently disabled;
    \item to serve in the armed forces (including those called to active duty);
    \item to serve with a foreign aid service of the Federal Government (e.g. Peace Corps);
    \item to serve on official church missions.
  }
}


\section{Student-to-Faculty Ratio}{
  Student-to-faculty ratio is defined as total
  FTE students (not in graduate or professional programs) divided by total
  FTE instructional staff (not teaching in graduate or professional programs).

}

\section{Reporting}{
All data on this file is applicable only to institutions with undergraduate
  students. Institutions with traditional academic year calendar systems
  (semester, quarter, trimester or 4-1-4) report their enrollment as of October
  15 or the official fall reporting date of the institution (academic year).
  Institutions with calendar systems that differ by program or allow continuous
  enrollment report students that are enrolled at any time between August 1 and
  October 31 (program year). Only academic year institutions report entering class
  data. Institutions with short programs (less-than-one-year) will not have data
  on retention.

  Beginning with the fall 2016 data, instititions
can include first-time bachelor’s-seeking (or equivalent) study abroad students
who were excluded from the first-time cohort but who have re-enrolled at the
institution their second year.

}

\section{More Info}{
\href{https://nces.ed.gov/ipeds/use-the-data/survey-components/8/fall-enrollment}{IPEDS Fall Enrollment Survey Components}
}
