% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{fct_case_when}
\alias{fct_case_when}
\title{Ordered Factor case_when()}
\usage{
fct_case_when(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas. See ?dplyr::case_when for details}
}
\value{
An ordered factor vector of length 1 or n, matching the length of the logical input or output vectors, with the type (and attributes) of the first RHS. Inconsistent lengths or types will generate an error.
}
\description{
\lifecycle{experimental}
Can replace `case_when()` syntax and outputs an ordered factor in the same order as the cases, useful for meaningful ordering in plots and tables.  This is because for `case_when()` the arguments are evaluated in order, so you must proceed from the most specific to the most general. Tables and plots will therefor be ordered by the evaluation order.
}
