% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tracer.R
\name{file_trace}
\alias{file_trace}
\title{Trace inputs and outputs for a file}
\usage{
file_trace(file, direction = "down", levels_up_max = 10, levels_down_max = 10)
}
\arguments{
\item{file}{The .R or .Rmd file you want to trace. Can be a partial match, extension not required.}

\item{direction}{A string either "down" (the default) or "right" specifying the direction the relationships are drawn.}

\item{levels_up_max}{\Sexpr[results=rd]{lifecycle::badge("experimental")} How many levels up the hierarchy to search and draw. Defaults to 10, values less than 2 are ignored.}

\item{levels_down_max}{\Sexpr[results=rd]{lifecycle::badge("experimental")} How many levels down the hierarchy to search and draw. Defaults to 10, values less than 3 are ignored.}
}
\value{
A connection graphic of code files and data files leading into the specified file, and resulting from the specified file.
}
\description{
\lifecycle{experimental}
Produces a connection graphic linking all inputs to a given file, and all outputs from the file.  Currently only works on code files.  The ability to specify a data file will be added.
}
\section{Note}{

This is a very early alpha release.  Please provide feedback and bug reports.

You will want to use the zoom window to view the result, and be aware that you can zoom in on it using the standard mouse actions, as some will be quite large and will appear impossible to read at the default zoom.
}

\examples{
\dontrun{
file_trace("institution_base")
}
}
