% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{freq_tibble}
\alias{freq_tibble}
\title{Generate a frequency tibble}
\usage{
freq_tibble(data, rows, cols, ...)
}
\arguments{
\item{rows}{The primary rows of the table (use groups for additional)}

\item{cols}{The columns of the table}

\item{...}{Additional grouping variables that will subdivide rows.}
}
\value{
A tibble
}
\description{
Generate a frequency table with marginal values
}
