\name{gr}
\alias{gr}

\title{Graduation Rates: GR}

\description{
This file contains the graduation rate status as of August 31 of the current year for the
cohort of full-time, first-time degree/certificate-seeking undergraduates
in both four year and two year institutions.
}

\section{Four-Year Institutions}{
Data for four year institutions include:
\itemize{
  \item the number of bachelor degree-seeking students who were enrolled 6 years prior (the cohort)
  \item the number of bachelor degree seeking students who completed any degree/
certificate within 150 percent of normal time
  \item the number who completed a bachelor's degree within 100, 125 or 150 percent of normal time
  \item the number of bachelor's degree-seeking students who transferred out.
  }

Data for students seeking a degree/certificate other than a bachelor's degree are also included
for four year institutions.
}

\section{Two-Year Institutions}{
Data for two year institutions include:
\itemize{
  \item the number of full-time, first-time students who were enrolled 3 years prior (the cohort).
  \item the number of students who completed any degree/certificate within 100 or 150 percent of normal
time.
  \item the number of students who transferred out.
  }
}


\section{Demographics}{
  Each record will contain the above data by gender and race/ethnicity. Note:
  The number of students who completed a degree/certificate within 100 percent of
  normal time is not available by race/ethnicity and gender.
}

\section{Glossary}{
\describe{
  \item{Time to completion}{The amount of time necessary for a student to complete all requirements for a degree or certificate according to the institution's catalog. This is typically 4 years (8 semesters or trimesters, or 12 quarters, excluding summer terms) for a bachelor's degree in a standard term-based institution; 2 years (4 semesters or trimesters, or 6 quarters, excluding summer terms) for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

  \item{Completers within 150\% (of normal time)}{Students (within a cohort or subcohort) who completed their program within 150\% of the normal time to completion.}
  \item{Bachelor's Degree}{An award (baccalaureate or equivalent degree, as determined by the Secretary, U.S. Department of Education) that normally requires at least 4 but NOT more than 5 years of full-time equivalent college-level work. This includes ALL bachelor's degrees conferred in a 5-year COOPERATIVE (WORK-STUDY PLAN) PROGRAM. A cooperative plan provides for alternate class attendance and employment in business, industry, or government; thus, it allows students to combine actual work experience with their college studies. Also, includes bachelor's degrees in which the normal 4 years of work are completed in 3 years.}
  \item{Total transfer-out students (non-completers)}{Total number of students from the cohort or subcohort who transferred out at any time within 150\% of normal time to completion. A transfer-out student is a student who has not completed or graduated from the program in which he or she was enrolled, but who has subsequently enrolled at another eligible institution in any program for which the reporting institution provided substantial preparation. }
  \item{Exclusions}{Students from the cohort or subcohort who left the institution for any of the following reasons: died or were totally and permanently disabled; to serve in the armed forces; to serve with a foreign aid service of the Federal Government, such as the Peace Corps; or to serve on official church missions."}
  }
}
