% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Value not matching (oppsoite of %in%)}
\usage{
... \%ni\% NA
}
\value{
A logical vector, indicating if there was no match for each element.
}
\description{
Returns a logical vector indicating where there is not a match for its left operation.  Opposite of %in%
}
