% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_error.R
\name{help_error_so}
\alias{help_error_so}
\title{Search Stack Overflow for help on errors}
\usage{
help_error_so(error = geterrmessage(), add_error_term = T)
}
\description{
Search the web for help on errors
}
