% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{install_project_files}
\alias{install_project_files}
\title{Install template files for SI to a project directory.}
\usage{
install_project_files(local_github_dir, overwrite = F)
}
\arguments{
\item{local_github_dir}{Where the cloned repo is.}

\item{overwrite}{If the files already exist, do you want to overwrite (default = FALSE)_}
}
\description{
Installs SI template files to project directory
}
\examples{
install_project_files()
}
