% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_add_info}
\alias{ipeds_add_info}
\title{Conveniently add school info to an IPEDS table.}
\usage{
ipeds_add_info(.data)
}
\arguments{
\item{.data}{An IPEDS table (must contain unitid and year).}
}
\value{
Original IPEDS table with school info added
}
\description{
\lifecycle{experimental}
Convenience function to add school info to an IPEDS table. Includes institution_entity_name, sector, and state.  More can be added if useful, contact JZ with suggestions.
}
\examples{
\dontrun{
efd <- efd \%>\% ipeds_add_info()
}
}
