% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_data}
\alias{ipeds_data}
\title{Table of cleaned IPEDs data}
\usage{
ipeds_data(.show_details = F)
}
\arguments{
\item{.show_details}{Include the path and the survey description? Defaults to FALSE for easier reading}
}
\value{
IPEDS data table
}
\description{
\lifecycle{experimental}
Shows a table of the cleaned IPEDs data.
}
\examples{
\dontrun{
ipeds_data()
}
}
