% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_info}
\alias{ipeds_info}
\title{Show IPEDS survey info}
\usage{
ipeds_info(survey_group)
}
\arguments{
\item{survey_group}{String value matching the abbreviated survey group}
}
\value{
Printed info about the survey
}
\description{
\lifecycle{experimental}
Shows the distinct overview and notes fields from the SI ipeds dictionary
}
\examples{
\dontrun{
ipeds_info("efc")
}
}
