% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_inst_lookup}
\alias{ipeds_inst_lookup}
\title{Show quick summary info on institution given a unitid}
\usage{
ipeds_inst_lookup(unitid)
}
\arguments{
\item{unitid}{Numeric `unitid value.}
}
\value{
Printed info about the institution.
}
\description{
Shows quick info on an institution.
}
\examples{
\dontrun{
ipeds_inst_lookup(230764)
}
}
