% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_unitid_lookup}
\alias{ipeds_unitid_lookup}
\title{Find unitids that match a string in the institution name}
\usage{
ipeds_unitid_lookup(instname)
}
\arguments{
\item{instname}{Full or partial string to detect in instution name.}
}
\value{
Printed info about the institution.
}
\description{
Show's unitid's for institution names that match the string provided
}
\examples{
\dontrun{
ipeds_unitid_lookup("Utah")
}
}
