% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_functions.R
\name{ipeds_visdat}
\alias{ipeds_visdat}
\title{vis_dat for ipeds}
\usage{
ipeds_visdat(.data, years = "all", .sample_frac = 0.1)
}
\arguments{
\item{years}{Single year or vector of years to plot.  Defaults to all years in data.}

\item{.sample_frac}{Percent of observations to sample from each year.  Defaults to .10.}
}
\value{
Plot of ipeds survey by year.
}
\description{
Produce a vis_dat plot for ipeds data split by year with optional sampling.
}
\examples{
\dontrun{
hd \%>\% ipeds_visdat(years = 2008:2011)
}
}
