% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{round_trailing_zeros}
\alias{round_trailing_zeros}
\title{Round a value and indicate how precise zeros are}
\usage{
round_trailing_zeros(x, digits = 2)
}
\arguments{
\item{x}{Numeric values to be rounded.}
}
\value{
Returns a character vector of rounded values including accurate zero values. Note that this rounds .5 to 1 (janitor::round_half_up) unlike base::round()
}
\description{
This function will take a value and round it.  If the rounded value becomes zero, trailing zeros are added to indicate how precise the 0 value is.
}
\examples{
# round_trailing_zeros(0.005, digits = 2)
}
