% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{SI_colorplot}
\alias{SI_colorplot}
\title{Show Sorenson Impact Theme Colors}
\usage{
SI_colorplot()
}
\value{
A plot of SI colors
}
\description{
Shows the Sorenson Impact theme colors for reference.
}
\examples{
SI_colorplot()
}
