% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SI_design}
\alias{SI_design}
\title{Sorenson Impact Custom Color Schemes}
\format{A named list of color hex values.
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
SI_design
}
\description{
A dataset containing a name listed of hex values used in SI themes
}
\keyword{datasets}
