% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{si_draft_full_report}
\alias{si_draft_full_report}
\title{Draft a Sorenson Impact Full Report}
\usage{
si_draft_full_report(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new Sorenson Impact fulll rmarkdown report
}
\examples{
draft_full_report("My presentation.Rmd")
}
