% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{si_draft_presentation}
\alias{si_draft_presentation}
\title{Draft a Sorenson Impact Presentation}
\usage{
si_draft_presentation(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new Sorenson Impact ioslides presentation.
}
\examples{
si_draft_presentation("My presentation.Rmd")
}
