% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{si_ggplot_theme_update}
\alias{si_ggplot_theme_update}
\title{Apply all Sorenson Impact ggplot themes}
\usage{
si_ggplot_theme_update()
}
\value{
Invisibly sets SI ggplot theme values.
}
\description{
Applies all Sorenson Impact custom colors and settings for ggplot
}
\examples{
\dontrun{
SI_ggplot_update()
}
}
