% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{SI_ggplot_update}
\alias{SI_ggplot_update}
\title{Apply all Sorenson Impact ggplot themes}
\usage{
SI_ggplot_update()
}
\value{
Invisibly sets SI ggplot theme values.
}
\description{
Applies all Sorenson Impact custom colors and settings for ggplot
}
\examples{
SI_colorplot()
}
