% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{si_ggsave}
\alias{si_ggsave}
\title{Save a branded plot}
\usage{
si_ggsave(
  filename = "auto",
  dir = "auto",
  plot = ggplot2::last_plot(),
  width = 6,
  height = 4,
  dpi = 300,
  add_logo = TRUE,
  logo_height_ratio = 0.05,
  band_color = si_design$granite
)
}
\arguments{
\item{filename}{Filename to create on disk. If "auto", defaults to the title of the \code{last_plot} within the \code{plot_directory}.}

\item{dir}{Directory to save file to.  Defaults to "auto", which places plots in a "plots" subdir of the script dir.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width}{Width in inches (default: 6).}

\item{height}{Height in inches (default: 4).}

\item{dpi}{Dots per inch. Defaults to 300.  The resolution of the file will be width*DPI by height*DPI.}

\item{add_logo}{Logical. Add the sorenson impact branding bar? (defaults: TRUE)}

\item{logo_height_ratio}{Number between 0 and 1, with sensibile values between .01 and .1.  The percent of the image height that the bar should be. Default is 0.05.}

\item{band_color}{The color of the SI Logo band.  Defaults to \code{si_design$granite}.}
}
\value{
A png of the last plot with optional SI logo band.
}
\description{
Save a \code{ggplot2} plot as png with an optional Sorenson Impact branding bar.
}
\examples{
\dontrun{
SI_colorplot() + ggplot2::ggtitle("My Title")
SI_ggsave(add_logo = TRUE)
}
}
