% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_google_drive_path_fix}
\alias{si_google_drive_path_fix}
\title{Check for correct google drive file stream path and correct if necessary.}
\usage{
si_google_drive_path_fix(check_only = F, osx_check_msg = T)
}
\arguments{
\item{check_only}{Tests if correct path exists but does not fix it (default: FALSE)}

\item{osx_check_msg}{Should a message be returned if not run on OSX. Defaults to TRUE. FALSE is used internal during package startup to generate warning.}
}
\value{
FALSE if path is not correct and check_only = TRUE. Otherwise informative message.
}
\description{
Checks for the existence of the correct google drive file stream path and optionally fix it.
\lifecycle{experimental}
}
\examples{
# si_google_drive_path_fix(check_only = T)
}
