% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_install_packages}
\alias{si_install_packages}
\title{Install all the packages we've ever found useful.}
\usage{
si_install_packages()
}
\description{
\lifecycle{experimental}
This is a convenience function for after R has been updated and all
the packages need to be reinstalled.
}
