% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{si_knitr_update}
\alias{si_knitr_update}
\title{Apply all SI knitr options}
\usage{
si_knitr_update()
}
\value{
Invisibly sets knitr options
}
\description{
Applies all Sorenson Impact custom knitr options.
}
\examples{
SI_knitr_update()
}
