% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_news}
\alias{si_news}
\title{Show news for new version}
\usage{
si_news(in_viewer = TRUE)
}
\arguments{
\item{in_viewer}{Logical. Show news in viewer instead of console (defaults to TRUE)}
}
\description{
\lifecycle{experimental}
Shows the news for the package
}
