% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_scale_big_dollar}
\alias{si_scale_big_dollar}
\title{Scale large dollar amounts to shortened dollar amounts (k, M, B, T)}
\usage{
si_scale_big_dollar(x, sep = " ", suffix_n = F)
}
\arguments{
\item{sep}{Seperator to use between number and unit (defaults to " ").}

\item{suffix_n}{Use "Bn" and "Tn" instead of "B" and "T".}
}
\value{
Character vector of formatted dollar values.
}
\description{
Format a vector of numeric values according to monetary abbreviations.

(ie: 1,000 becomes "1 k", 2,500,000 becomes "2.5 B") See: https://www.wallstreetoasis.com/forums/abbreviation-for-thousands-millions-billions-trillion
}
\examples{
si_scale_big_dollar(1000)
si_scale_big_dollar(1000000000)
si_scale_big_dollar(1000000000, suffix_n = T)
}
