% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_update}
\alias{si_update}
\title{Update the sorensonimpact package}
\usage{
si_update()
}
\description{
\lifecycle{experimental}
Automatically unloads, updates, and reloads the sorensonimpact package.
}
