% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visdat_custom_functions.R
\name{si_visdat_grouped}
\alias{si_visdat_grouped}
\title{vis_dat for grouped data}
\usage{
si_visdat_grouped(.data, ..., method = "vis_dat", .sample_frac = "auto")
}
\arguments{
\item{...}{bare, unquoted column(s) to use as the index to group by. Alternatively will accept a grouped df.}

\item{method}{Which visdat function to use. One of "vis_dat", "vis_miss", or "vis_value".  Accepts shorthand "dat", "val", and "miss".}

\item{.sample_frac}{Percent of observations to sample from each year.  Default "auto" samples down to 100,000 rows, split evenly between groups for vis_dat. For vis_miss and vis_value, "auto" uses all data.}
}
\value{
visdat plot separated by grouping variable.
}
\description{
\lifecycle{experimental}
Produce a vis_dat plot for ipeds data split by year with optional sampling.
}
\examples{
\dontrun{
diamonds \%>\% si_visdat_group(facet_group = cut)
}
}
