% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensonimpact-package.R
\docType{package}
\name{sorensonimpact}
\alias{sorensonimpact}
\title{sorensonimpact}
\description{
This package contains internal functions used in Sorenson Impact Center code.
}
\keyword{internal}
