% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{sum_rowwise}
\alias{sum_rowwise}
\title{Sum selected columns by row}
\usage{
sum_rowwise(data, ..., sum_col = "sum", na.rm = FALSE)
}
\arguments{
\item{...}{Columns to sum.}

\item{sum_col}{Name of sum column. Defaults to "sum".}

\item{na.rm}{Remove NAs? Passed to rowSums}
}
\value{
Vector with rowwise sums.
}
\description{
Sum selected columns within mutate without \code{rowwise()} (which can be very slow).
}
\examples{
cars \%>\% sum_rowwise(speed, dist, na.rm = T, sum_col = "mysum"))
}
