% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{tp}
\alias{tp}
\title{Tibble Preview}
\usage{
tp(data, rows = 10)
}
\value{
A preview of a tibble.
}
\description{
\lifecycle{experimental}
Show a sample of all tibble data without hiding columns.
}
