% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{update_si}
\alias{update_si}
\title{Update the sorensonimpact package}
\usage{
update_si()
}
\description{
\lifecycle{experimental}
Automatically unloads, updates, and reloads the sorensonimpact package.
}
