% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{view_df}
\alias{view_df}
\title{View selected data frame}
\usage{
view_df()
}
\description{
The RStudio Environment pane variable name column is too narrow for long
names, so difficutl to find the right data frame in long list of similar
names. Select the variable and use shortcut to use data viewer on it.
}
\details{
Stolen from: https://github.com/dracodoc/mischelper/blob/master/R/misc.R
}
