% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xview.R
\name{xview}
\alias{xview}
\title{Show a table in Excel}
\usage{
xview(.data)
}
\arguments{
\item{.data}{A table.}
}
\description{
\lifecycle{experimental}
Opens the table in a temporary file in excel.
}
\examples{
\dontrun{
diamonds \%>\% xview()
}
}
