% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tracer.R
\name{file_trace}
\alias{file_trace}
\title{Trace inputs and outputs for a file}
\usage{
file_trace(
  file,
  code_path = "project",
  trim_data_path = NULL,
  direction = "down",
  levels_up_max = 10,
  levels_down_max = 10
)
}
\arguments{
\item{file}{The .R or .Rmd file you want to trace. Can be a partial match for code file, extension not required. For data file, full file name is required with extension.}

\item{code_path}{The top-level path to search for .R and .Rmd files. The search is recursive to cover all child directories. Defaults to "project" which will search the active project directory.}

\item{trim_data_path}{Optional string to remove from the final displayed output for data file nodes. If most of the data files are in a common directory and thus the information is not informative, specifying that directory will clean up the output by shortening the full path to only the relative path.}

\item{direction}{A string either "down" (the default) or "right" specifying the direction the relationships are drawn.}

\item{levels_up_max}{\Sexpr[results=rd]{lifecycle::badge("experimental")} How many levels up the hierarchy to search and draw. Defaults to 10, values less than 2 are ignored.}

\item{levels_down_max}{\Sexpr[results=rd]{lifecycle::badge("experimental")} How many levels down the hierarchy to search and draw. Defaults to 10, values less than 3 are ignored.}
}
\value{
A connection graphic of code files and data files leading into the specified file, and resulting from the specified file.
}
\description{
\lifecycle{experimental}
Produces a connection graphic linking all inputs to a given file, and all outputs from the file.  Currently only works on code files.  The ability to specify a data file will be added.
}
\section{Note}{

This is a very early alpha release.  Please provide feedback and bug reports.

You will want to use the zoom window to view the result, and be aware that you can zoom in on it using the standard mouse actions, as some will be quite large and will appear impossible to read at the default zoom.
}

\examples{
\dontrun{
file_trace("institution_base")
}
}
