% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{si_colorplot}
\alias{si_colorplot}
\title{Show Sorenson Impact Theme Colors}
\usage{
si_colorplot()
}
\value{
A plot of SI colors
}
\description{
Shows the Sorenson Impact theme colors for reference.
}
\examples{
\dontrun{
SI_colorplot()
}
}
