% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_design}
\alias{si_design}
\title{Sorenson Impact Custom Color Schemes}
\format{
A named list of color hex values.
}
\usage{
si_design
}
\description{
A dataset containing a name listed of hex values used in SI themes
}
\keyword{datasets}
