% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{si_news}
\alias{si_news}
\title{Show news for new version}
\usage{
si_news(all = FALSE)
}
\arguments{
\item{all}{Logical. Show all previous news in viewer. Defaults to FALSE.}
}
\description{
\lifecycle{experimental}
Shows the news for the package
}
